feat(refactor-mcts): Separate InvalidChoice class from ChooseFromNumbers
- Add comments for doxygen
